rename module and added remove style on chenge effects
