+++ js/effects.js
@@ -95,6 +95,7 @@ function effectChangeHandler (evt) {
 
 const resetSlider = () => {
   sliderElement.noUiSlider.updateOptions(EFFECT_CONFIG.none);
+  uploadImageElement.removeAttribute('style');
   hideSlider();
 };
 
@@ -113,6 +114,7 @@ sliderElement.noUiSlider.on('update', () => {
   const type = uploadImageElement.dataset.type;
 
   if (type === Effect.NONE) {
+    uploadImageElement.removeAttribute('style');
     return hideSlider();
   }
