+++ js/main.js
@@ -2,7 +2,7 @@ import {openPicture} from './picture-modal.js';
 import {getItemById} from './utils.js';
 import {renderPictures} from './pictures.js';
 import {getData} from './api.js';
-import {showAlert} from './message.js';
+import {showAlert} from './messages.js';
 import './upload-form.js';
 
 const picturesElement = document.querySelector('.pictures');
